change tool name
