+++ .github/workflows/meta-data-check.yml
@@ -18,4 +18,4 @@ jobs:
       - name: Validate metadata
         run: |
           pip install -U solar-registry
-          solar-registry validate unittest
+          solar-registry validate python_unittest
