+++ python_unittest/script/linux/install.sh
@@ -8,7 +8,7 @@ echo "${TOOL_ROOT}"
 cd "${TOOL_ROOT}"
 
 # 安装到全局的python中，无需创建虚拟环境(当前已经使用uniSDK，对运行环境的依赖很少)
-pip install -r requirements.txt
+pip install "testsolar-testtool-sdk>=0.2.1" "pydantic>=2" "loguru"
 
 # 安装uniSDK
 curl -Lk https://testsolar-1321258242.cos.ap-guangzhou.myqcloud.com/cli/testtools_sdk-install/stable/install.sh | bash
