+++ python_unittest/script/linux/nodeCleanup.sh
@@ -1,0 +1,3 @@
+#! /bin/bash
+
+set -exu -o pipefail
