+++ python_unittest/script/linux/nodeSetup.sh
@@ -1,0 +1,3 @@
+#! /bin/bash
+
+set -exu -o pipefail
