+++ python_unittest/script/linux/run.sh
@@ -8,4 +8,4 @@ echo "${TOOL_ROOT}"
 echo "${TESTSOLAR_WORKSPACE}"
 
 /usr/local/bin/testtools_sdk version
-/usr/local/bin/testtools_sdk serve --tool unittest
+/usr/local/bin/testtools_sdk serve --tool python_unittest
