--- unittest/requirements.txt
@@ -1,13 +1,0 @@
-# This file is @generated by PDM.
-# Please do not edit it manually.
-
-annotated-types==0.7.0
-dacite==1.8.1
-loguru==0.7.2
-lxml==5.2.2
-portalocker==2.10.0
-pydantic==2.8.2
-pydantic-core==2.20.1
-testsolar-testtool-sdk==0.3.2
-typing-extensions==4.12.2
-unittest-xml-reporting==3.2.0
